+++ src/tensor/algstrct.cxx
@@ -352,7 +352,7 @@ namespace CTF_int {
     printf("CTF ERROR: cannot convert elements of this algebraic structure to CSR\n");
     ASSERT(0);
   }
-  void algstrct::coo_to_ccsr(int64_t nz, int64_t nnz_row, char * csr_vs, int * csr_cs, int * csr_rs, char const * coo_vs, int const * coo_rs, int const * coo_cs) const {
+  void algstrct::coo_to_ccsr(int64_t nz, int64_t nnz_row, char * csr_vs, int * csr_cs, int * csr_rs, char const * coo_vs, int64_t const * coo_rs, int64_t const * coo_cs) const {
     printf("CTF ERROR: cannot convert elements of this algebraic structure to CCSR\n");
     ASSERT(0);
   }
@@ -363,7 +363,7 @@ namespace CTF_int {
     ASSERT(0);
   }
 
-  void algstrct::ccsr_to_coo(int64_t nz, int64_t nnz_row, char const * csr_vs, int const * csr_ja, int const * csr_ia, int const * row_enc, char * coo_vs, int * coo_rs, int * coo_cs) const {
+  void algstrct::ccsr_to_coo(int64_t nz, int64_t nnz_row, char const * csr_vs, int const * csr_ja, int const * csr_ia, int const * row_enc, char * coo_vs, int64_t * coo_rs, int64_t * coo_cs) const {
     printf("CTF ERROR: cannot convert elements of this algebraic structure to CCSR\n");
     ASSERT(0);
   }
@@ -740,19 +740,19 @@ namespace CTF_int {
     printf("CTF ERROR: csrmm not present for this algebraic structure\n");
     ASSERT(0);
   }
-  void algstrct::ccsrmm(int64_t      m,
-                        int64_t      n,
-                        int64_t      k,
-                        int64_t      nnz_row,
-                        char const * alpha,
-                        char const * A,
-                        int const *  JA,
-                        int const *  IA,
-                        int const *  row_enc,
-                        int64_t      nnz_A,
-                        char const * B,
-                        char const * beta,
-                        char *&      C,
+  void algstrct::ccsrmm(int64_t         m,
+                        int64_t         n,
+                        int64_t         k,
+                        int64_t         nnz_row,
+                        char const *    alpha,
+                        char const *    A,
+                        int const *     JA,
+                        int const *     IA,
+                        int64_t const * row_enc,
+                        int64_t         nnz_A,
+                        char const *    B,
+                        char const *    beta,
+                        char *&         C,
                         CTF_int::bivar_function const * func) const {
     ASSERT(0);
   }
